refactor(group): extract normalizeUrl utility
